Move db update check out of StandardIncludes to its own listener+controller
